maybe building on mac

+++ nui/Unlogger.cpp
@@ -69,7 +69,7 @@ void Unlogger::process() {
         }
       }
 
-      if (abs(tc-last_elapsed) > 50e6) {
+      if (abs(((long long)tc-(long long)last_elapsed)) > 50e6) {
         //qDebug() << "elapsed";
         emit elapsed();
         last_elapsed = tc;
